WIP intelligent table parsing from .docx files

+++ scripts/etc/normalisescripts/getcharacters.py
@@ -22,15 +22,16 @@ def split_characters(names):
     return collect
 
 
-def process(paths, ext, out, prefix):
+def process(paths, ext, out, prefix, dry_run):
     for p in paths:
         name = os.path.basename(p).split('.')[0]
         out_path = os.path.join(out, f'{prefix}_{name}_out.names')
-        with open(out_path, 'w') as file:
-            names = sorted(set(split_characters(adr.get_column_data(p, 3))))
-            for n in names:
-                file.write(f'{n}\n')
-            file.close()
+        names = sorted(set(split_characters(adr.get_column_data(p, 3))))
+        if not dry_run:
+            with open(out_path, 'w') as file:
+                for n in names:
+                    file.write(f'{n}\n')
+                file.close()
 
 
 def group_items(items, size):
@@ -75,6 +76,8 @@ def main():
                         help='write to file can be a or w')
     parser.add_argument('--process-count', type=int, nargs='?', default=4,
                         help='total processes to spawn in pool; cannot be higher than system total')
+    parser.add_argument('--dry-run', action='store_true',
+                        help='perform a dry run')
     args = parser.parse_args()
 
     write_type = args.write_type.lower()
@@ -98,7 +101,7 @@ def main():
 
     pool = []
     for i, p in enumerate(grouped_paths):
-        proc = mp.Process(target=process, args=(p, args.ext, out_path, f'process{i}'))
+        proc = mp.Process(target=process, args=(p, args.ext, out_path, f'process{i}', args.dry_run))
         pool.append(proc)
 
     for p in pool:
